% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbilling_objects.R
\name{BillingAccount}
\alias{BillingAccount}
\title{Google Cloud Billing API Objects 
Retrieves Google Developers Console billing accounts and associates them with projects.}
\usage{
BillingAccount(name = NULL, open = NULL, displayName = NULL)
}
\arguments{
\item{name}{The resource name of the billing account}

\item{open}{True if the billing account is open, and will therefore be charged for any usage on associated projects}

\item{displayName}{The display name given to the billing account, such as `My Billing Account`}
}
\value{
BillingAccount object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:56:43
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudbillingv1.auto/R/cloudbilling_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BillingAccount Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A billing account in [Google Developers Console](https://console.developers.google.com/). You can assign a billing account to one or more projects.
}
