% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbilling_functions.R
\name{billingAccounts.get}
\alias{billingAccounts.get}
\title{Gets information about a billing account. The current authenticated user must be an [owner of the billing account](https://support.google.com/cloud/answer/4430947).}
\usage{
billingAccounts.get(name)
}
\arguments{
\item{name}{The resource name of the billing account to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/billing/}{Google Documentation}
}

