% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbilling_functions.R
\name{billingAccounts.list}
\alias{billingAccounts.list}
\title{Lists the billing accounts that the current authenticated user [owns](https://support.google.com/cloud/answer/4430947).}
\usage{
billingAccounts.list(pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{pageSize}{Requested page size}

\item{pageToken}{A token identifying a page of results to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/billing/}{Google Documentation}
}
