% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbilling_objects.R
\name{ListBillingAccountsResponse}
\alias{ListBillingAccountsResponse}
\title{ListBillingAccountsResponse Object}
\usage{
ListBillingAccountsResponse(billingAccounts = NULL, nextPageToken = NULL)
}
\arguments{
\item{billingAccounts}{A list of billing accounts}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
ListBillingAccountsResponse object
}
\description{
ListBillingAccountsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for `ListBillingAccounts`.
}

