% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbilling_objects.R
\name{ListProjectBillingInfoResponse}
\alias{ListProjectBillingInfoResponse}
\title{ListProjectBillingInfoResponse Object}
\usage{
ListProjectBillingInfoResponse(projectBillingInfo = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{projectBillingInfo}{A list of `ProjectBillingInfo` resources representing the projects associated with the billing account}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
ListProjectBillingInfoResponse object
}
\description{
ListProjectBillingInfoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for `ListProjectBillingInfoResponse`.
}

