% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbilling_objects.R
\name{ProjectBillingInfo}
\alias{ProjectBillingInfo}
\title{ProjectBillingInfo Object}
\usage{
ProjectBillingInfo(name = NULL, projectId = NULL,
  billingAccountName = NULL, billingEnabled = NULL)
}
\arguments{
\item{name}{The resource name for the `ProjectBillingInfo`; has the form `projects/{project_id}/billingInfo`}

\item{projectId}{The ID of the project that this `ProjectBillingInfo` represents, such as `tokyo-rain-123`}

\item{billingAccountName}{The resource name of the billing account associated with the project, if any}

\item{billingEnabled}{True if the project is associated with an open billing account, to which usage on the project is charged}
}
\value{
ProjectBillingInfo object
}
\description{
ProjectBillingInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encapsulation of billing information for a Developers Console project. A project has at most one associated billing account at a time (but a billing account can be assigned to multiple projects).
}
\seealso{
Other ProjectBillingInfo functions: \code{\link{projects.updateBillingInfo}}
}

