% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbilling_functions.R
\name{projects.getBillingInfo}
\alias{projects.getBillingInfo}
\title{Gets the billing information for a project. The current authenticated user must have [permission to view the project](https://cloud.google.com/docs/permissions-overview#h.bgs0oxofvnoo ).}
\usage{
projects.getBillingInfo(name)
}
\arguments{
\item{name}{The resource name of the project for which billing information is retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/billing/}{Google Documentation}
}

